further refactored server file so that passport is in auth directory and only routes are in routes.
